refactor: reduce duplicated code based on SonarQube advice
